backend/requirements.txt
asgiref==3.5.2
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
python-dotenv==0.20.0
sqlparse==0.4.2

hazardest/game/dist/game-0.1.tar.gz
